refactor: searchTarget 타입 변경, 폴더구조 변경
코트 검색을 위해 searchTarget의 타입 변경,
data.tsx코드가 너무 복잡해서 코드 분리
